envoy: extract helpers to generate types
