+++ pkg/generator/generator.go
@@ -5,7 +5,6 @@ import (
 	"kourier/pkg/config"
 	"kourier/pkg/envoy"
 	"kourier/pkg/knative"
-	"net/http"
 	"os"
 	"strconv"
 	"time"
@@ -15,11 +14,8 @@ import (
 	"knative.dev/pkg/tracker"
 
 	v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
 	endpoint "github.com/envoyproxy/go-control-plane/envoy/api/v2/endpoint"
 	route "github.com/envoyproxy/go-control-plane/envoy/api/v2/route"
-	"github.com/golang/protobuf/ptypes"
-	"github.com/golang/protobuf/ptypes/wrappers"
 	log "github.com/sirupsen/logrus"
 	kubev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -131,11 +127,11 @@ func addIngressToCaches(caches *Caches,
 				publicLbEndpoints := lbEndpointsForKubeEndpoints(endpoints, targetPort)
 
 				connectTimeout := 5 * time.Second
-				cluster := clusterForRevision(split.ServiceName, connectTimeout, publicLbEndpoints, http2, path)
+				cluster := envoy.NewCluster(split.ServiceName+path, connectTimeout, publicLbEndpoints, http2)
 
 				caches.AddCluster(&cluster, split.ServiceName, split.ServiceNamespace, path)
 
-				weightedCluster := weightedCluster(split.ServiceName, uint32(split.Percent), path, headersSplit)
+				weightedCluster := envoy.NewWeightedCluster(split.ServiceName+path, uint32(split.Percent), headersSplit)
 
 				wrs = append(wrs, &weightedCluster)
 			}
@@ -148,19 +144,11 @@ func addIngressToCaches(caches *Caches,
 		}
 
 		externalDomains := knative.ExternalDomains(&rule, localDomainName)
-		virtualHost := route.VirtualHost{
-			Name:    ingress.Name,
-			Domains: externalDomains,
-			Routes:  ruleRoute,
-		}
+		virtualHost := envoy.NewVirtualHost(ingress.Name, externalDomains, ruleRoute)
 
 		// External should also be accessible internally
 		internalDomains := append(knative.InternalDomains(&rule, localDomainName), externalDomains...)
-		internalVirtualHost := route.VirtualHost{
-			Name:    ingress.Name,
-			Domains: internalDomains,
-			Routes:  ruleRoute,
-		}
+		internalVirtualHost := envoy.NewVirtualHost(ingress.Name, internalDomains, ruleRoute)
 
 		if knative.RuleIsExternal(&rule, ingress.GetSpec().Visibility) {
 			externalVirtualHosts = append(externalVirtualHosts, &virtualHost)
@@ -202,15 +190,14 @@ func listenersFromVirtualHosts(externalVirtualHosts []*route.VirtualHost,
 }
 
 func internalKourierVirtualHost(ikrs []*route.Route) route.VirtualHost {
-	return route.VirtualHost{
-		Name:    config.InternalKourierDomain,
-		Domains: []string{config.InternalKourierDomain},
-		Routes:  ikrs,
-	}
+	return envoy.NewVirtualHost(
+		config.InternalKourierDomain,
+		[]string{config.InternalKourierDomain},
+		ikrs,
+	)
 }
 
 func internalKourierRoutes(ingresses []*v1alpha1.Ingress) []*route.Route {
-
 	var hashes []string
 	var routes []*route.Route
 	for _, ingress := range ingresses {
@@ -223,64 +210,25 @@ func internalKourierRoutes(ingresses []*v1alpha1.Ingress) []*route.Route {
 	}
 
 	for _, hash := range hashes {
-		r := route.Route{
-			Name: fmt.Sprintf("%s_%s", config.InternalKourierDomain, hash),
-			Match: &route.RouteMatch{
-				PathSpecifier: &route.RouteMatch_Path{
-					Path: fmt.Sprintf("%s/%s", config.InternalKourierPath, hash),
-				},
-			},
-			Action: &route.Route_DirectResponse{
-				DirectResponse: &route.DirectResponseAction{Status: http.StatusOK},
-			},
-		}
+		name := fmt.Sprintf("%s_%s", config.InternalKourierDomain, hash)
+		path := fmt.Sprintf("%s/%s", config.InternalKourierPath, hash)
+		r := envoy.NewRouteStatusOK(name, path)
 		routes = append(routes, &r)
 	}
 
-	staticRoute := route.Route{
-		Name: config.InternalKourierDomain,
-		Match: &route.RouteMatch{
-			PathSpecifier: &route.RouteMatch_Path{
-				Path: config.InternalKourierPath,
-			},
-		},
-		Action: &route.Route_DirectResponse{
-			DirectResponse: &route.DirectResponseAction{Status: http.StatusOK},
-		},
-	}
-
+	staticRoute := envoy.NewRouteStatusOK(
+		config.InternalKourierDomain,
+		config.InternalKourierPath,
+	)
 	routes = append(routes, &staticRoute)
 
 	return routes
 }
 
 func lbEndpointsForKubeEndpoints(kubeEndpoints *kubev1.Endpoints, targetPort int32) (publicLbEndpoints []*endpoint.LbEndpoint) {
-
 	for _, subset := range kubeEndpoints.Subsets {
-
 		for _, address := range subset.Addresses {
-
-			serviceEndpoint := &core.Address{
-				Address: &core.Address_SocketAddress{
-					SocketAddress: &core.SocketAddress{
-						Protocol: core.SocketAddress_TCP,
-						Address:  address.IP,
-						PortSpecifier: &core.SocketAddress_PortValue{
-							PortValue: uint32(targetPort),
-						},
-						Ipv4Compat: true,
-					},
-				},
-			}
-
-			lbEndpoint := endpoint.LbEndpoint{
-				HostIdentifier: &endpoint.LbEndpoint_Endpoint{
-					Endpoint: &endpoint.Endpoint{
-						Address: serviceEndpoint,
-					},
-				},
-			}
-
+			lbEndpoint := envoy.NewLBEndpoint(address.IP, uint32(targetPort))
 			publicLbEndpoints = append(publicLbEndpoints, &lbEndpoint)
 		}
 	}
@@ -289,6 +237,8 @@ func lbEndpointsForKubeEndpoints(kubeEndpoints *kubev1.Endpoints, targetPort int
 }
 
 func createRouteForRevision(routeName string, i int, httpPath *v1alpha1.HTTPIngressPath, wrs []*route.WeightedCluster_ClusterWeight) route.Route {
+	name := routeName + "_" + strconv.Itoa(i)
+
 	path := "/"
 	if httpPath.Path != "" {
 		path = httpPath.Path
@@ -299,64 +249,6 @@ func createRouteForRevision(routeName string, i int, httpPath *v1alpha1.HTTPIngr
 		routeTimeout = httpPath.Timeout.Duration
 	}
 
-	r := route.Route{
-		Name: routeName + "_" + strconv.Itoa(i),
-		Match: &route.RouteMatch{
-			PathSpecifier: &route.RouteMatch_Prefix{
-				Prefix: path,
-			},
-		},
-		Action: &route.Route_Route{Route: &route.RouteAction{
-			ClusterSpecifier: &route.RouteAction_WeightedClusters{
-				WeightedClusters: &route.WeightedCluster{
-					Clusters: wrs,
-				},
-			},
-			Timeout: ptypes.DurationProto(routeTimeout),
-			UpgradeConfigs: []*route.RouteAction_UpgradeConfig{{
-				UpgradeType: "websocket",
-				Enabled:     &wrappers.BoolValue{Value: true},
-			}},
-			RetryPolicy: createRetryPolicyForRoute(httpPath),
-		}},
-		RequestHeadersToAdd: headersToAdd(httpPath.AppendHeaders),
-	}
-
-	return r
-}
-
-func weightedCluster(revisionName string, trafficPerc uint32, path string, headers map[string]string) route.WeightedCluster_ClusterWeight {
-	return route.WeightedCluster_ClusterWeight{
-		Name: revisionName + path,
-		Weight: &wrappers.UInt32Value{
-			Value: trafficPerc,
-		},
-		RequestHeadersToAdd: headersToAdd(headers),
-	}
-}
-
-func headersToAdd(headers map[string]string) []*core.HeaderValueOption {
-	var res []*core.HeaderValueOption
-
-	for headerName, headerVal := range headers {
-		header := core.HeaderValueOption{
-			Header: &core.HeaderValue{
-				Key:   headerName,
-				Value: headerVal,
-			},
-			Append: &wrappers.BoolValue{
-				Value: true,
-			},
-		}
-
-		res = append(res, &header)
-
-	}
-
-	return res
-}
-
-func createRetryPolicyForRoute(httpPath *v1alpha1.HTTPIngressPath) *route.RetryPolicy {
 	attempts := 0
 	var perTryTimeout time.Duration
 	if httpPath.Retries != nil {
@@ -367,43 +259,9 @@ func createRetryPolicyForRoute(httpPath *v1alpha1.HTTPIngressPath) *route.RetryP
 		}
 	}
 
-	if attempts > 0 {
-		return &route.RetryPolicy{
-			RetryOn: "5xx",
-			NumRetries: &wrappers.UInt32Value{
-				Value: uint32(attempts),
-			},
-			PerTryTimeout: ptypes.DurationProto(perTryTimeout),
-		}
-	} else {
-		return nil
-	}
-}
-
-func clusterForRevision(revisionName string, connectTimeout time.Duration, publicLbEndpoints []*endpoint.LbEndpoint, http2 bool, path string) v2.Cluster {
-
-	cluster := v2.Cluster{
-		Name: revisionName + path,
-		ClusterDiscoveryType: &v2.Cluster_Type{
-			Type: v2.Cluster_STRICT_DNS,
-		},
-		ConnectTimeout: ptypes.DurationProto(connectTimeout),
-		LoadAssignment: &v2.ClusterLoadAssignment{
-			ClusterName: revisionName + path,
-			Endpoints: []*endpoint.LocalityLbEndpoints{
-				{
-					LbEndpoints: publicLbEndpoints,
-					Priority:    1,
-				},
-			},
-		},
-	}
-
-	if http2 {
-		cluster.Http2ProtocolOptions = &core.Http2ProtocolOptions{}
-	}
-
-	return cluster
+	return envoy.NewRoute(
+		name, path, wrs, routeTimeout, uint32(attempts), perTryTimeout, httpPath.AppendHeaders,
+	)
 }
 
 func useHTTPSListener() bool {
